requirements.txt
streamlit==1.41.1
google-generativeai==0.4.1
PyPDF2==3.0.1
python-dotenv==1.0.1
faiss-cpu==1.7.4
huggingface-hub==0.11.1
sentence-transformers==2.2.2
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
tqdm==4.67.1
numpy==1.23.5
scipy==1.14.1
typing-extensions==4.12.2
InstructorEmbedding
